Gfx: Move Texture into Gfx
